+++ src/widgets/clock/variants/defaultClock/validateParams.test.ts
@@ -18,13 +18,74 @@
 import { describe, expect, test } from 'vitest';
 import { validateParams } from './validateParams';
 
-test('success when params are undefined', () => {
-  const { isValid, expectedArgument } = validateParams();
-  expect(isValid).toBe(true);
-  expect(expectedArgument).toBe(undefined);
+describe('success', () => {
+  test('when params is undefined', () => {
+    const { isValid, expectedArgument } = validateParams();
+    expect(isValid).toBe(true);
+    expect(expectedArgument).toBe(undefined);
+  });
+
+  test('when params.dateFormat is a string', () => {
+    const testParams = { dateFormat: 'YY-MM-DD' };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(true);
+    expect(expectedArgument).toBe(undefined);
+  });
+
+  test('when params.timeFormat is a string', () => {
+    const testParams = { timeFormat: 'h:mm:ss' };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(true);
+    expect(expectedArgument).toBe(undefined);
+  });
+
+  test('when params.dateFormat and params.timeFormat are strings', () => {
+    const testParams = {
+      dateFormat: 'YY-MM-DD',
+      timeFormat: 'h:mm:ss',
+    };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(true);
+    expect(expectedArgument).toBe(undefined);
+  });
 });
 
-describe('failure when params is defined', () => {
+describe('failure when params is an object and has invalid keys', () => {
+  test('dateFormat is not a string', () => {
+    const testParams = {
+      dateFormat: 1,
+    };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(false);
+    expect(expectedArgument?.length).gt(0);
+  });
+
+  test('timeFormat is not a string', () => {
+    const testParams = {
+      timeFormat: 1,
+    };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(false);
+    expect(expectedArgument?.length).gt(0);
+  });
+
+  test('an extra key is present', () => {
+    const testParams = {
+      blarp: '1',
+    };
+
+    const { isValid, expectedArgument } = validateParams(testParams);
+    expect(isValid).toBe(false);
+    expect(expectedArgument?.length).gt(0);
+  });
+});
+
+describe('failure when params is defined and not an object', () => {
   test('params is a number', () => {
     const { isValid, expectedArgument } = validateParams(1);
     expect(isValid).toBe(false);